+++ mall-spring-boot/src/main/java/edu/nju/mall/repository/OrderRepository.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @Author: qianen.yin
  * @CreateDate: 2020-02-03 14:03
  */
-public interface OrderRepository extends JpaRepository<Order, Integer>, JpaSpecificationExecutor<Order> {
+public interface OrderRepository extends JpaRepository<Order, Long>, JpaSpecificationExecutor<Order> {
 }
